Fixed the specs to fit previous split.
