+++ spec/spec_helper.rb
@@ -4,6 +4,13 @@ require "rubygems"
 #RSpec.configure do |config|
 #end
 
+def silence_warnings
+  old_verbose, $VERBOSE = $VERBOSE, nil
+  yield
+ensure
+  $VERBOSE = old_verbose
+end
+
 def reset_i18n()
   I18n.backend = I18n::Backend::Simple.new
 end
